// src/app/movies/api/movie-info.ts
import { Movies } from "./movies";
export interface MoviesInfo{
    results: MovieInfo[]
}
export interface MovieInfo {
    adult:                 boolean;
    backdrop_path:         string;
    belongs_to_collection: null;
    budget:                number;
    genres:                Genre[];
    homepage:              string;
    id:                    number;
    imdb_id:               string;
    origin_country:        string[];
    original_language:     string;
    original_title:        string;
    overview:              string;
    popularity:            number;
    poster_path:           string;
    production_companies:  ProductionCompany[];
    production_countries:  ProductionCountry[];
    release_date:          Date;
    revenue:               number;
    runtime:               number;
    spoken_languages:      SpokenLanguage[];
    status:                string;
    tagline:               string;
    title:                 string;
    video:                 boolean;
    vote_average:          number;
    vote_count:            number;
    videos:                Videos;
    images:                Images;
    credits:               Credits;
    similar:               Movies;
    recommendations:       Movies;
}

export interface Credits {
    cast: Cast[];
    crew: Cast[];
}

export interface Cast {
    adult:                boolean;
    gender:               number;
    id:                   number;
    known_for_department: Department;
    name:                 string;
    original_name:        string;
    popularity:           number;
    profile_path:         null | string;
    cast_id?:             number;
    character?:           string;
    credit_id:            string;
    order?:               number;
    department?:          Department;
    job?:                 string;
}

export enum Department {
    Acting = "Acting",
    Camera = "Camera",
    Directing = "Directing",
    Editing = "Editing",
    Production = "Production",
    Sound = "Sound",
    Writing = "Writing",
}

export interface Genre {
    id:   number;
    name: string;
}

export interface Images {
    backdrops: any[];
    logos:     any[];
    posters:   any[];
}

export interface ProductionCompany {
    id:             number;
    logo_path:      null | string;
    name:           string;
    origin_country: string;
}

export interface ProductionCountry {
    iso_3166_1: string;
    name:       string;
}

// export interface Similar {
//     page:          number;
//     results:       SimilarResult[];
//     total_pages:   number;
//     total_results: number;
// }

// export interface SimilarResult {
//     adult:             boolean;
//     backdrop_path:     null | string;
//     genre_ids:         number[];
//     id:                number;
//     original_language: string;
//     original_title:    string;
//     overview:          string;
//     popularity:        number;
//     poster_path:       string;
//     release_date:      Date;
//     title:             string;
//     video:             boolean;
//     vote_average:      number;
//     vote_count:        number;
// }

export interface SpokenLanguage {
    english_name: string;
    iso_639_1:    string;
    name:         string;
}

export interface Videos {
    results: VideosResult[];
}

export interface VideosResult {
    iso_639_1:    string;
    iso_3166_1:   string;
    name:         string;
    key:          string;
    published_at: Date;
    site:         string;
    size:         number;
    type:         string;
    official:     boolean;
    id:           string;
}